Implemented two versions of discounting.
Strategy pattern which injects the calculation method into the order

Decorator pattern which creates a chain of discounts which each use the
output from the previous discount decorator as the input for itself.

+++ src/ordersystem/interfaces/OrderTotalCalculator.java
@@ -1,10 +1,9 @@
 package ordersystem.interfaces;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import ordersystem.OrderLine;
 
 public interface OrderTotalCalculator {
-	BigDecimal getTotal(List<OrderLine> orderLines); 
+	float getTotal(List<OrderLine> orderLines); 
 }
